Restructured db and patched the auth bug
