+++ template/build.gradle
@@ -4,7 +4,7 @@ plugins {
 	id 'io.spring.dependency-management' version '1.1.0'
 }
 
-group = 'pot.insurence'
+group = 'pot.insurance'
 version = '0.0.1-SNAPSHOT'
 sourceCompatibility = '17'
