+++ template/src/main/java/pot/insurance/manager/ManagerApplication.java
@@ -1,4 +1,4 @@
-package pot.insurence.manager;
+package pot.insurance.manager;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
