+++ template/src/main/java/pot/insurance/manager/controller/EchoController.java
@@ -1,4 +1,4 @@
-package pot.insurence.manager.controller;
+package pot.insurance.manager.controller;
 
 import java.security.Principal;
 import org.springframework.web.bind.annotation.GetMapping;
