+++ template/src/main/resources/db/changelog/changes/db_schema.sql
@@ -1,0 +1,5 @@
+-- liquibase formatted sql
+
+-- changeset dovias:1
+CREATE SCHEMA `java_internship`;
+-- rollback DROP SCHEMA `java_internship`;
