+++ template/src/test/java/pot/insurance/manager/ManagerApplicationTests.java
@@ -1,4 +1,4 @@
-package pot.insurence.manager;
+package pot.insurance.manager;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
